add availableOperands query
