+++ lib/api/operands/index.ts
@@ -1,0 +1,5 @@
+import RandomInt from "./RandomInt";
+
+export default [
+  RandomInt
+]
