+++ lib/graphql/resolvers/availableOperands.ts
@@ -1,0 +1,6 @@
+import { RandomOperand } from "../../api/Problem";
+import availableOperands from "../../api/operands/index";
+
+export default {
+  availableOperands: (): RandomOperand[] => availableOperands
+}
