+++ lib/graphql/resolvers/index.ts
@@ -1,7 +1,9 @@
 import helloWorld from "./helloWorld";
 import problems from "./problems";
+import availableOperands from "./availableOperands";
 
 export default {
   ...helloWorld,
-  ...problems
+  ...problems,
+  ...availableOperands
 }
